+++ src/ts/ui/initUI.ts
@@ -1,6 +1,7 @@
 import {i18n} from "../i18n";
 import {html2md} from "../sv/html2md";
 import {setEditMode} from "../toolbar/EditMode";
+import {toggleOutline} from "../toolbar/Outline";
 import {setContentTheme} from "./setContentTheme";
 import {setTheme} from "./setTheme";
 
@@ -75,6 +76,9 @@ export const initUI = (vditor: IVditor) => {
     }
 
     setEditMode(vditor, vditor.options.mode, afterRender(vditor, contentElement));
+    if (vditor.options.outline && vditor.toolbar.elements.outline) {
+        toggleOutline(vditor, vditor.toolbar.elements.outline.firstElementChild);
+    }
 };
 
 export const setPadding = (vditor: IVditor) => {