+++ src/ts/markdown/previewRender.ts
@@ -78,6 +78,13 @@ export const md2html = (mdText: string, options?: IPreviewOptions) => {
             setext: mergedOptions.markdown.setext,
             toc: mergedOptions.markdown.toc,
         });
+        if (options.renderers) {
+            lute.SetJSRenderers({
+                renderers: {
+                    Md2HTML: options.renderers,
+                },
+            });
+        }
         return lute.Md2HTML(mdText);
     });
 };