+++ CHANGELOG.md
@@ -60,6 +60,7 @@
 ### v3.1.19 / 2020-04-2x
 
 * [324](https://github.com/Vanessa219/vditor/issues/324) 支持多款主题预览 `引入特性`
+* [341](https://github.com/Vanessa219/vditor/issues/341) Can not delete the image at IR mode `修复缺陷`
 
 ### v3.1.18 / 2020-04-24