+++ CHANGELOG.md
@@ -56,11 +56,11 @@
 ### TODO
 
 * [open issues](https://github.com/Vanessa219/vditor/issues)
+* [324](https://github.com/Vanessa219/vditor/issues/324) 支持多款主题预览 `引入特性`
 
 ### v3.1.19 / 2020-04-2x
 
 * [325](https://github.com/Vanessa219/vditor/issues/325) 导出功能 `引入特性`
-* [324](https://github.com/Vanessa219/vditor/issues/324) 支持多款主题预览 `引入特性`
 * [341](https://github.com/Vanessa219/vditor/issues/341) Can not delete the image at IR mode `修复缺陷`
 
 ### v3.1.18 / 2020-04-24